crypto-ml/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic[email]>=2.0.0
python-multipart>=0.0.6
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
requests>=2.25.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0
psycopg2-binary>=2.9.0
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
